# docs/src/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "ord",
    "description": "◉ Ordinal wallet and block explorer",
    "license": {
      "name": "CC0-1.0"
    },
    "version": "0.8.2"
  },
  "paths": {
    "/api/v1/brc20/address/{address}/balance": {
      "get": {
        "tags": [
          "brc20"
        ],
        "operationId": "get all ticker balances of the address",
        "parameters": [
          {
            "name": "address",
            "in": "path",
            "description": "Address",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Obtain account balances by query address.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BRC20AllBalance"
                }
              }
            }
          },
          "400": {
            "description": "Bad query.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                },
                "example": {
                  "code": 2,
                  "msg": "bad request"
                }
              }
            }
          },
          "404": {
            "description": "Not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                },
                "example": {
                  "code": 3,
                  "msg": "not found"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                },
                "example": {
                  "code": 1,
                  "msg": "internal error"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/brc20/tick/{ticker}/address/{address}/balance": {
      "get": {
        "tags": [
          "brc20"
        ],
        "operationId": "get the ticker balance of the address",
        "parameters": [
          {
            "name": "ticker",
            "in": "path",
            "description": "Token ticker",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "address",
            "in": "path",
            "description": "Address",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Obtain account balance by query ticker.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BRC20Balance"
                }
              }
            }
          },
          "400": {
            "description": "Bad query.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                },
                "example": {
                  "code": 2,
                  "msg": "bad request"
                }
              }
            }
          },
          "404": {
            "description": "Not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                },
                "example": {
                  "code": 3,
                  "msg": "not found"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                },
                "example": {
                  "code": 1,
                  "msg": "internal error"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/brc20s/address/{address}/balance": {
      "get": {
        "tags": [
          "brc20s"
        ],
        "operationId": "get all ticker balances of the address",
        "parameters": [
          {
            "name": "address",
            "in": "path",
            "description": "Address",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Obtain account balances by query address.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BRC20SAllBalance"
                }
              }
            }
          },
          "400": {
            "description": "Bad query.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                },
                "example": {
                  "code": 2,
                  "msg": "bad request"
                }
              }
            }
          },
          "404": {
            "description": "Not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                },
                "example": {
                  "code": 3,
                  "msg": "not found"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                },
                "example": {
                  "code": 1,
                  "msg": "internal error"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/brc20s/tick/{tick_id}/address/{address}/balance": {
      "get": {
        "tags": [
          "brc20s"
        ],
        "operationId": "get the ticker balance of the address",
        "parameters": [
          {
            "name": "tick_id",
            "in": "path",
            "description": "Token ticker ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "address",
            "in": "path",
            "description": "Address",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Obtain account balance by query ticker.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BRC20SBalance"
                }
              }
            }
          },
          "400": {
            "description": "Bad query.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                },
                "example": {
                  "code": 2,
                  "msg": "bad request"
                }
              }
            }
          },
          "404": {
            "description": "Not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                },
                "example": {
                  "code": 3,
                  "msg": "not found"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                },
                "example": {
                  "code": 1,
                  "msg": "internal error"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ApiError": {
        "oneOf": [
          {
            "type": "object",
            "required": [
              "Internal"
            ],
            "properties": {
              "Internal": {
                "type": "string",
                "description": "Internal server error."
              }
            },
            "example": {
              "code": 1,
              "msg": "internal error"
            }
          },
          {
            "type": "object",
            "required": [
              "BadRequest"
            ],
            "properties": {
              "BadRequest": {
                "type": "string",
                "description": "Bad request."
              }
            },
            "example": {
              "code": 1,
              "msg": "bad request"
            }
          },
          {
            "type": "object",
            "required": [
              "NotFound"
            ],
            "properties": {
              "NotFound": {
                "type": "string",
                "description": "Resource not found."
              }
            },
            "example": {
              "code": 1,
              "msg": "not found"
            }
          }
        ]
      },
      "BRC20AllBalance": {
        "type": "object",
        "required": [
          "code",
          "msg",
          "data"
        ],
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32"
          },
          "data": {
            "$ref": "#/components/schemas/brc20.AllBalance"
          },
          "msg": {
            "type": "string",
            "description": "ok",
            "example": "ok"
          }
        }
      },
...
      "BRC20Balance": {
        "type": "object",
        "required": [
          "code",
          "msg",
          "data"
        ],
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32"
          },
          "data": {
            "$ref": "#/components/schemas/brc20.Balance"
          },
          "msg": {
            "type": "string",
            "description": "ok",
            "example": "ok"
          }
        }
      },
...
      "BRC20SAllBalance": {
        "type": "object",
        "required": [
          "code",
          "msg",
          "data"
        ],
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32"
          },
          "data": {
            "$ref": "#/components/schemas/brc20s.AllBalance"
          },
          "msg": {
            "type": "string",
            "description": "ok",
            "example": "ok"
          }
        }
      },
...
      "BRC20SBalance": {
        "type": "object",
        "required": [
          "code",
          "msg",
          "data"
        ],
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32"
          },
          "data": {
            "$ref": "#/components/schemas/brc20s.Balance"
          },
          "msg": {
            "type": "string",
            "description": "ok",
            "example": "ok"
          }
        }
      },
...
      "brc20.AllBalance": {
        "type": "object",
        "required": [
          "balance"
        ],
        "properties": {
          "balance": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/brc20.Balance"
            }
          }
        }
      },
...
      "brc20.Balance": {
        "type": "object",
        "required": [
          "tick",
          "availableBalance",
          "transferableBalance",
          "overallBalance"
        ],
        "properties": {
          "availableBalance": {
            "type": "string",
            "format": "uint64",
            "description": "Available balance."
          },
          "overallBalance": {
            "type": "string",
            "format": "uint64",
            "description": "Overall balance."
          },
          "tick": {
            "type": "string",
            "description": "Name of the ticker."
          },
          "transferableBalance": {
            "type": "string",
            "format": "uint64",
            "description": "Transferable balance."
          }
        }
      },
...
      "brc20s.AllBalance": {
        "type": "object",
        "required": [
          "balance"
        ],
        "properties": {
          "balance": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/brc20s.Balance"
            }
          }
        }
      },
...
      "brc20s.Balance": {
        "type": "object",
        "required": [
          "tick",
          "transferable",
          "overall"
        ],
        "properties": {
          "overall": {
            "type": "string",
            "format": "uint64",
            "description": "Overall balance."
          },
          "tick": {
            "$ref": "#/components/schemas/brc20s.Tick"
          },
          "transferable": {
            "type": "string",
            "format": "uint64",
            "description": "Transferable balance."
          }
        }
      },
...
      "brc20s.Tick": {
        "type": "object",
        "required": [
          "id",
          "name"
        ],
        "properties": {
          "id": {
            "type": "string",
            "description": "Id of the ticker."
          },
          "name": {
            "type": "string",
            "description": "Name of the ticker."
          }
        }
      },
...
    }
  }
}